better naming of some vars and moving some files into a 'utils/' folder
